gpt-executor/requirements.txt
fastapi
uvicorn
pandas
numpy
openpyxl
matplotlib
seaborn
PyPDF2
pdfplumber
pytesseract
pdf2image
Pillow
opencv-python-headless
requests
beautifulsoup4
lxml
PyMuPDF
python-docx
pyyaml
xmltodict
python-multipart